"Joe" Directory that includes the first webpage using p5 to create a button with text
The most important files are:
- sketch.js
- index.html

sketch.js is responsible for drawing, while index.html is the actual webpage that can be displayed. If you are in VS code, you can right click index.html and click "Open with live server" to see the page displayed on your browser.

If you want a simpler view, go to https://editor.p5js.org/ and just paste the sketch.js code so that you can play around with it yourself.